Story 2599. Renamed CadcMastResolver to OldCadcMastResolver. The changes in this story should stay in this branch and should not be merged to the master branch.
