+++ caom2-artifact-resolvers/build.gradle
@@ -15,7 +15,7 @@ sourceCompatibility = 1.7
 
 group = 'org.opencadc'
 
-version = '1.1.14'
+version = '1.1.15'
 
 dependencies {
     compile 'log4j:log4j:[1.2.0,)'
@@ -58,7 +58,7 @@ tasks.withType(Test) {
 
 task intTest(type: Test) {
     // set the configuration context
-    testClassesDir = sourceSets.intTest.output.classesDir
+    testClassesDirs = sourceSets.intTest.output.classesDirs
     classpath = sourceSets.intTest.runtimeClasspath
 
     // run the tests always
